requirements.txt
openai>=1.0.0
pyyaml>=6.0
pydantic>=2.0.0
rich>=13.0.0
numpy>=1.24.0
sentence-transformers>=3.0.0
pytest>=8.0.0
